Bug 513790 - Compiler warnings in I20170316-2000 after moving to M6 compiler

+++ bundles/org.eclipse.swt/Eclipse SWT WebKit/gtk/org/eclipse/swt/browser/WebKit.java
@@ -1207,7 +1207,7 @@ private static class Webkit2JavascriptEvaluator {
 			usedCallbackIds.add(value);
 			return value;
 		}
-		private static void removeId(long id) {
+		private static void removeId(int id) {
 			usedCallbackIds.remove(id);
 		}
 	}
